finish product table data
