+++ lib/br_invoices_pdf.rb
@@ -49,5 +49,6 @@ require 'br_invoices_pdf/util/mount_params'
 require 'br_invoices_pdf/util/enum'
 require 'br_invoices_pdf/util/base_renderer'
 require 'br_invoices_pdf/util/pdf_renderer'
+require 'br_invoices_pdf/util/product_table'
 require 'br_invoices_pdf/cfe'
 require 'br_invoices_pdf/nfce'
