+++ lib/br_invoices_pdf/util/base_renderer.rb
@@ -29,22 +29,6 @@ module BrInvoicesPdf
         end
       end
 
-      PRODUCT_TABLE_BASE_DATA = [['CÓD.', 'DESCRIÇÃO', 'QTD.', 'UND.', 'V.UNIT', 'V.TOT']].freeze
-      # :reek:FeatureEnvy
-      def product_table_data(data)
-        data[:products].reduce(PRODUCT_TABLE_BASE_DATA) do |result, cur|
-          result + [[
-            cur[:code],
-            cur[:description],
-            format_number(cur[:quantity]),
-            cur[:unit_label],
-            format_currency(cur[:unit_value]),
-            format_currency(cur[:total_value])
-          ]]
-        end
-      end
-      private_class_method :product_table_data
-
       CNPJ_FORMAT = '%02d.%03d.%03d/%04d-%02d'
       # :reek:FeatureEnvy
       def format_cnpj(cnpj)
